[FLINK-1874] [streaming] Connectors separated into individual projects
