+++ flink-staging/flink-streaming/flink-streaming-connectors/flink-kafka-connector/src/main/java/org/apache/flink/streaming/connectors/kafka/partitioner/SerializableKafkaPartitioner.java
