+++ flink-staging/flink-streaming/flink-streaming-connectors/flink-kafka-connector/src/test/resources/log4j-test.properties
