+++ flink-staging/flink-streaming/flink-streaming-connectors/flink-rabbitmq-connector/pom.xml
@@ -25,21 +25,35 @@ under the License.
 
 	<parent>
 		<groupId>org.apache.flink</groupId>
-		<artifactId>flink-streaming-parent</artifactId>
+		<artifactId>flink-streaming-connectors-parent</artifactId>
 		<version>0.9-SNAPSHOT</version>
 		<relativePath>..</relativePath>
 	</parent>
 
-	<artifactId>flink-streaming-connectors-parent</artifactId>
-	<name>flink-streaming-connectors</name>
+	<artifactId>flink-rabbitmq-connector</artifactId>
+	<name>flink-rabbitmq-connector</name>
 
-	<packaging>pom</packaging>
+	<packaging>jar</packaging>
 
-	<modules>
-		<module>flink-flume-connector</module>
-		<module>flink-kafka-connector</module>
-		<module>flink-rabbitmq-connector</module>
-		<module>flink-twitter-connector</module>
-	</modules>
+	<!-- Allow users to pass custom connector versions -->
+	<properties>
+		<rabbitmq.version>3.3.1</rabbitmq.version>
+	</properties>
+
+	<dependencies>
+
+		<dependency>
+			<groupId>org.apache.flink</groupId>
+			<artifactId>flink-streaming-core</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>com.rabbitmq</groupId>
+			<artifactId>amqp-client</artifactId>
+			<version>${rabbitmq.version}</version>
+		</dependency>
+
+	</dependencies>
 
 </project>
