+++ flink-staging/flink-streaming/flink-streaming-connectors/flink-rabbitmq-connector/src/main/java/org/apache/flink/streaming/connectors/rabbitmq/RMQSink.java
