+++ flink-staging/flink-streaming/flink-streaming-connectors/flink-twitter-connector/src/main/java/org/apache/flink/streaming/connectors/twitter/TwitterStreaming.java
